Moved Login component to Auth folder, updated imports
